Upgraded Build Machinery

+++ .circleci/builder/service.py
@@ -7,7 +7,7 @@ import typing
 from builder.constants import BUILD_BASE_DIR, ARTIFACT_DEST_DIR
 from builder.utils import filter_gitignore_entry__as_string, Collection, BuildJob, Category, Notebook, load_gitignore_data, run_command
 
-def build_categories(start_path: str, begin_path: str = None) -> types.GeneratorType:
+def build_categories(start_path: str, rebuild: bool = True) -> types.GeneratorType:
     gitignore_data = load_gitignore_data(os.path.join(start_path, '.gitignore'))
     for root, dirnames, filenames in os.walk(start_path):
         for dirname in dirnames:
@@ -35,26 +35,26 @@ def build_categories(start_path: str, begin_path: str = None) -> types.Generator
                     raise NotImplementedError(f'Category missing Requirements File[{requirements_path}]')
 
                 build_dir = os.path.join(BUILD_BASE_DIR, dirname)
-                if os.path.exists(build_dir):
+                if os.path.exists(build_dir) and rebuild:
                     shutil.rmtree(build_dir)
 
                 artifact_dir = os.path.join(ARTIFACT_DEST_DIR, dirname)
-                if os.path.exists(artifact_dir):
+                if os.path.exists(artifact_dir) and rebuild:
                     shutil.rmtree(artifact_dir)
 
                 yield Category(dirname, notebooks, dirpath, build_dir, artifact_dir)
 
             else:
-                for category in build_categories(dirpath, start_path):
+                for category in build_categories(dirpath, rebuild):
                     yield category
 
-def find_collections(notebook_collection_paths: typing.List[str]) -> types.GeneratorType:
+def find_collections(notebook_collection_paths: typing.List[str], rebuild: bool = True) -> types.GeneratorType:
     for name in notebook_collection_paths:
         c_path = os.path.join(os.getcwd(), name)
-        yield Collection(name, [cate for cate in build_categories(c_path)])
+        yield Collection(name, [cate for cate in build_categories(c_path, rebuild)])
 
-def find_build_jobs(notebook_collection_paths: typing.List[str]):
-    for collection in find_collections(notebook_collection_paths):
+def find_build_jobs(notebook_collection_paths: typing.List[str], rebuild: bool = True):
+    for collection in find_collections(notebook_collection_paths, rebuild):
         for category in collection.categories:
             # category.setup_build_env()
             # category.inject_extra_files()
@@ -71,7 +71,22 @@ def setup_build(job: BuildJob) -> None:
     for notebook in job.category.notebooks:
         notebook.create_build_script([job.collection.name, job.category.name], job.category.build_dir, job.category.artifact_dir)
 
+def build_website(job: BuildJob) -> None:
+    import pdb; pdb.set_trace()
+# def consolidate_artifacts(job: BuildJob) -> None:
+#     dest_dir = os.path.join(ARTIFACT_DEST_DIR, 'html-files')
+#     if not os.path.exists(dest_dir):
+#         os.makedirs(dest_dir)
+# 
+#     for notebook in job.category.notebooks:
+#         filename = notebook.filename.rsplit('.', 1)[0]
+#         filename = f'{filename}.html'
+#         filepath = f'{ARTIFACT_DEST_DIR}/{job.collection.name}/{job.category.name}/{filename}'
+#         dest_filepath = os.path.join(dest_dir, filename)
+#         shutil.copyfile(filepath, dest_filepath)
+
 def run_build(job: BuildJob) -> None:
     for script in job.scripts:
         command = f'bash "{script}"'
         run_command(command)
+
